Upgrade Next.js to 15, React to 19, and related deps
- next: ^14 → ^15 (frontpage, addon-catalog, ui)
- react/react-dom: ^18 → ^19 (all workspace packages)
- framer-motion: ^11 → ^12 (fix AnimationControls type, 'pop'→'spring', RefObject null)
- next-plausible: ^3 → ^4 (add src param, simplify PlausibleProvider)
- @storybook/nextjs: ^8.1.1 → ^8.6
- Await async params in releases, recipes, docs slug pages
- Await headers() in versions route and docs footer actions
- Extract getLlmsBannerLines to lib/ (Next.js 15 disallows non-route exports)
- Remove JSX.Element return types (removed from @types/react@19)
- Add legacy-peer-deps=true to .npmrc
