+++ apps/addon-catalog/package.json
@@ -20,16 +20,16 @@
     "@storybook/icons": "^1.2.9",
     "@types/mdx": "^2.0.13",
     "copy-to-clipboard": "^3.3.3",
-    "framer-motion": "^11.2.12",
+    "framer-motion": "^12",
     "graphql-request": "^7.1.0",
     "human-format": "^1.2.0",
-    "next": "^14.2.4",
-    "next-plausible": "^3.12.0",
+    "next": "^15",
+    "next-plausible": "^4.0.0",
     "next-themes": "^0.3.0",
     "posthog-js": "^1.306.0",
     "prismjs": "^1.29.0",
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "^19",
+    "react-dom": "^19",
     "rehype": "^13.0.1",
     "rehype-stringify": "^10.0.0",
     "rehype-urls": "^1.2.0",
@@ -41,8 +41,8 @@
   "devDependencies": {
     "@types/node": "^20",
     "@types/prismjs": "^1.26.4",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
     "eslint": "^8",
     "eslint-config-next": "14.2.4",
     "postcss": "^8",
