+++ apps/frontpage/app/layout.tsx
@@ -59,7 +59,7 @@ export default function RootLayout({
   children,
 }: {
   children: React.ReactNode;
-}): JSX.Element {
+}) {
   return (
     <html lang="en" suppressHydrationWarning>
       <body
