+++ apps/frontpage/package.json
@@ -22,7 +22,7 @@
     "@google-cloud/bigquery": "^7.7.0",
     "@mdx-js/loader": "^3.0.1",
     "@mdx-js/react": "^3.0.1",
-    "@next/bundle-analyzer": "^14.2.5",
+    "@next/bundle-analyzer": "^15",
     "@next/mdx": "^14.2.4",
     "@next/third-parties": "^14.2.4",
     "@radix-ui/react-accordion": "^1.1.2",
@@ -39,20 +39,20 @@
     "cookies-next": "^4.1.1",
     "copy-to-clipboard": "^3.3.3",
     "dedent": "^1.5.3",
-    "framer-motion": "^11.3.2",
+    "framer-motion": "^12",
     "fs-extra": "^11.2.0",
     "graphql-request": "^7.1.0",
     "gray-matter": "^4.0.3",
     "image-size": "^1.1.1",
-    "next": "^14.2.5",
+    "next": "^15",
     "next-mdx-remote": "^5.0.0",
-    "next-plausible": "^3.12.0",
+    "next-plausible": "^4.0.0",
     "next-themes": "^0.3.0",
     "node-fetch": "^3.3.2",
     "posthog-js": "^1.306.0",
     "posthog-node": "^5.17.2",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "^19",
+    "react-dom": "^19",
     "react-lite-youtube-embed": "^2.4.0",
     "react-syntax-highlighter": "^15.5.0",
     "react-use-scroll-direction": "^0.1.0",
@@ -69,7 +69,7 @@
   },
   "devDependencies": {
     "@chromatic-com/storybook": "^1.4.0",
-    "@next/eslint-plugin-next": "^14.1.1",
+    "@next/eslint-plugin-next": "^15",
     "@repo/eslint-config": "*",
     "@repo/tailwind-config": "*",
     "@repo/typescript-config": "*",
@@ -77,7 +77,7 @@
     "@storybook/addon-interactions": "^8.1.1",
     "@storybook/addon-links": "^8.1.1",
     "@storybook/blocks": "^8.1.1",
-    "@storybook/nextjs": "^8.1.1",
+    "@storybook/nextjs": "^8.6",
     "@storybook/preview-api": "^8.1.1",
     "@storybook/react": "^8.1.1",
     "@storybook/test": "^8.1.1",
@@ -85,8 +85,8 @@
     "@types/fs-extra": "^11.0.4",
     "@types/node": "^20",
     "@types/prompts": "^2.4.9",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
     "@types/react-syntax-highlighter": "^15.5.12",
     "@types/tar": "^6.1.13",
     "@vitejs/plugin-react": "^4.3.0",
