+++ packages/ui/package.json
@@ -30,9 +30,9 @@
     "copy-to-clipboard": "^3.3.3",
     "formik": "^2.4.6",
     "he": "^1.2.0",
-    "next": "^14.2.3",
+    "next": "^15",
     "next-mdx-remote": "^5.0.0",
-    "next-plausible": "^3.12.0",
+    "next-plausible": "^4.0.0",
     "posthog-js": "^1.306.0",
     "rehype-pretty-code": "^0.13.2",
     "rehype-slug": "^6.0.0",
@@ -56,13 +56,13 @@
     "@types/eslint": "^8.56.5",
     "@types/he": "^1.2.3",
     "@types/node": "^20.11.24",
-    "@types/react": "^18.2.61",
-    "@types/react-dom": "^18.2.19",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
     "autoprefixer": "^10.4.18",
     "eslint": "^8.57.0",
     "eslint-plugin-storybook": "^0.11.1",
     "postcss": "^8.4.35",
-    "react": "^18.2.0",
+    "react": "^19",
     "storybook": "^8.4.5",
     "tailwindcss": "^3.4.1",
     "typescript": "^5.3.3"
