+++ packages/utils/package.json
@@ -15,10 +15,10 @@
     "@turbo/gen": "^1.12.4",
     "@types/eslint": "^8.56.5",
     "@types/node": "^20.11.24",
-    "@types/react": "^18.2.61",
-    "@types/react-dom": "^18.2.19",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
     "eslint": "^8.57.0",
-    "react": "^18.2.0",
+    "react": "^19",
     "typescript": "^5.3.3"
   },
   "dependencies": {
